# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KEYFORTA API
  version: 1.0.0
  description: Versioned contract for the multi-organization rental and property operations platform.
servers:
  - url: https://api.keyforta.com/api/v1
  - url: http://localhost:3000/api/v1
security:
  - bearerAuth: []
paths:
...
  /viewing-requests:
    post:
      tags: [Public]
      operationId: requestViewing
      security: []
      requestBody: { $ref: '#/components/requestBodies/ViewingRequest' }
      responses: { '202': { $ref: '#/components/responses/PublicRequestAccepted' }, '429': { $ref: '#/components/responses/RateLimited' } }
...
  /landlord-onboarding-applications:
    post:
      tags: [Identity]
      operationId: submitLandlordOnboardingApplication
      requestBody: { $ref: '#/components/requestBodies/LandlordOnboardingApplication' }
      responses: { '201': { $ref: '#/components/responses/LandlordOnboardingApplication' }, '400': { $ref: '#/components/responses/Problem' }, '401': { $ref: '#/components/responses/Problem' }, '409': { $ref: '#/components/responses/Problem' }, '503': { $ref: '#/components/responses/Problem' } }
    get:
      tags: [Administration]
      operationId: listLandlordOnboardingApplications
      description: Requires a verified Entra oid in the configured platform-admin allowlist. Denied identities receive a generic not-found response.
      responses: { '200': { $ref: '#/components/responses/LandlordOnboardingApplicationList' }, '401': { $ref: '#/components/responses/Problem' }, '404': { $ref: '#/components/responses/Problem' }, '503': { $ref: '#/components/responses/Problem' } }
  /landlord-onboarding-applications/{applicationId}/decision:
    post:
      tags: [Administration]
      operationId: decideLandlordOnboardingApplication
      description: Requires a verified Entra oid in the configured platform-admin allowlist.
      parameters: [{ $ref: '#/components/parameters/LandlordOnboardingApplicationId' }]
      requestBody: { $ref: '#/components/requestBodies/LandlordOnboardingDecision' }
      responses: { '200': { $ref: '#/components/responses/LandlordOnboardingApplication' }, '400': { $ref: '#/components/responses/Problem' }, '401': { $ref: '#/components/responses/Problem' }, '404': { $ref: '#/components/responses/Problem' }, '409': { $ref: '#/components/responses/Problem' }, '503': { $ref: '#/components/responses/Problem' } }
...
  /properties:
    get: { tags: [Inventory], operationId: listProperties, security: [], parameters: [{ $ref: '#/components/parameters/PublicPropertyCity' }, { $ref: '#/components/parameters/PublicPropertyDistrict' }, { $ref: '#/components/parameters/PublicPropertyMinBedrooms' }, { $ref: '#/components/parameters/PublicPropertyMaxRent' }, { $ref: '#/components/parameters/Pagination' }, { $ref: '#/components/parameters/PublicPropertyLimit' }, { $ref: '#/components/parameters/PublicPropertySort' }], responses: { '200': { $ref: '#/components/responses/PublicPropertyList' } } }
    post: { tags: [Inventory], operationId: createProperty, requestBody: { $ref: '#/components/requestBodies/Property' }, responses: { '201': { $ref: '#/components/responses/Property' } } }
  /properties/{propertyId}:
    get: { tags: [Inventory], operationId: getProperty, security: [], parameters: [{ $ref: '#/components/parameters/PublicPropertyId' }], responses: { '200': { $ref: '#/components/responses/PublicProperty' } } }
    patch: { tags: [Inventory], operationId: updateProperty, parameters: [{ $ref: '#/components/parameters/PropertyId' }, { $ref: '#/components/parameters/IfMatch' }], requestBody: { $ref: '#/components/requestBodies/Patch' }, responses: { '200': { $ref: '#/components/responses/Property' } } }
    delete: { tags: [Inventory], operationId: archiveProperty, parameters: [{ $ref: '#/components/parameters/PropertyId' }], responses: { '204': { description: Archived } } }
...
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  parameters:
    Pagination: { name: cursor, in: query, schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$' } }
    PublicPropertyCity: { name: city, in: query, schema: { type: string, minLength: 1, maxLength: 120 } }
    PublicPropertyDistrict: { name: district, in: query, schema: { type: string, minLength: 1, maxLength: 120 } }
    PublicPropertyMinBedrooms: { name: minBedrooms, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
    PublicPropertyMaxRent: { name: maxMonthlyRentMinor, in: query, description: Maximum monthly rent in integer minor units encoded as a decimal string, schema: { type: string, pattern: '^[1-9][0-9]{0,17}$' } }
    PublicPropertyLimit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    PublicPropertySort: { name: sort, in: query, schema: { type: string, enum: [created_at_desc, name_asc, name_desc], default: created_at_desc } }
    PublicPropertyId: { name: propertyId, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$' } }
...
    LandlordOnboardingApplicationId: { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IfMatch: { name: If-Match, in: header, schema: { type: string } }
...
    PropertyId: { name: propertyId, in: path, required: true, schema: { type: string } }
...
  requestBodies:
...
    ViewingRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ViewingRequestInput' } } } }
...
    LandlordOnboardingApplication: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LandlordOnboardingApplicationInput' } } } }
    LandlordOnboardingDecision: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LandlordOnboardingDecisionInput' } } } }
...
    Property: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PropertyInput' } } } }
...
    Patch: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  responses:
    PublicRequestAccepted: { description: Request accepted for processing, content: { application/json: { schema: { $ref: '#/components/schemas/PublicRequestReceipt' } } } }
    RateLimited: { description: Caller exceeded the viewing-request rate limit, headers: { Retry-After: { schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    Property: { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
    PublicProperty: { description: Published public-safe property, content: { application/json: { schema: { $ref: '#/components/schemas/PublicPropertyEnvelope' } } } }
    PublicPropertyList: { description: Published public-safe properties, content: { application/json: { schema: { $ref: '#/components/schemas/PublicPropertyListEnvelope' } } } }
...
    LandlordOnboardingApplication: { description: Landlord onboarding application, content: { application/json: { schema: { $ref: '#/components/schemas/LandlordOnboardingApplicationEnvelope' } } } }
    LandlordOnboardingApplicationList: { description: Landlord onboarding review queue, content: { application/json: { schema: { $ref: '#/components/schemas/LandlordOnboardingApplicationListEnvelope' } } } }
    Problem: { description: Correlated error response, content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
    Problem: { type: object, required: [error], properties: { error: { type: object, required: [code, details, message, traceId], properties: { code: { type: string }, details: { type: object }, message: { type: string }, traceId: { type: string } }, additionalProperties: false } }, additionalProperties: false }
    PublicRequestReceipt: { type: object, required: [data, meta], properties: { data: { type: object, required: [reference, status], properties: { reference: { type: string }, status: { type: string, enum: [accepted] } }, additionalProperties: false }, meta: { $ref: '#/components/schemas/Meta' } }, additionalProperties: false }
...
    ViewingRequestInput: { type: object, required: [propertyId, name, email], properties: { propertyId: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$' }, name: { type: string, minLength: 2, maxLength: 120 }, email: { type: string, format: email, maxLength: 254 }, phone: { type: string, minLength: 7, maxLength: 40 }, preferredAt: { type: string, format: date-time, pattern: 'Z$' }, message: { type: string, maxLength: 1000 }, locale: { type: string, enum: [en, fr] }, website: { type: string, maxLength: 0, description: Honeypot field; must remain empty } }, additionalProperties: false }
...
    Meta: { type: object, required: [requestId], properties: { requestId: { type: string } }, additionalProperties: false }
...
    PublicPropertyProjection: { type: object, required: [id, name, address, city, district, summary, bedrooms, bathrooms, monthlyRentMinor, currency, availableFrom, amenities, imageUrls], properties: { id: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$' }, name: { type: string }, address: { type: string, description: Approved approximate location only }, city: { type: string }, district: { type: string }, summary: { type: string }, bedrooms: { type: integer, minimum: 0 }, bathrooms: { type: integer, minimum: 1 }, areaSquareMeters: { type: integer, minimum: 1 }, monthlyRentMinor: { type: string, pattern: '^[1-9][0-9]*$' }, currency: { type: string, pattern: '^[A-Z]{3}$' }, availableFrom: { type: string, format: date }, amenities: { type: array, items: { type: string } }, imageUrl: { type: string }, imageUrls: { type: array, minItems: 1, items: { type: string } } }, additionalProperties: false }
    PublicPropertyEnvelope: { type: object, required: [data, meta], properties: { data: { $ref: '#/components/schemas/PublicPropertyProjection' }, meta: { $ref: '#/components/schemas/Meta' } }, additionalProperties: false }
    PublicPropertyListEnvelope: { type: object, required: [items, total, nextCursor, meta], properties: { items: { type: array, items: { $ref: '#/components/schemas/PublicPropertyProjection' } }, total: { type: integer, minimum: 0 }, nextCursor: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$', nullable: true }, meta: { $ref: '#/components/schemas/Meta' } }, additionalProperties: false }
...
    LandlordOnboardingApplicationInput: { type: object, required: [applicantName, proposedOrganizationName], properties: { applicantName: { type: string, minLength: 2, maxLength: 120 }, proposedOrganizationName: { type: string, minLength: 2, maxLength: 160 } }, additionalProperties: false }
    LandlordOnboardingDecisionInput: { type: object, required: [decision, reason], properties: { decision: { type: string, enum: [approved, rejected] }, reason: { type: string, minLength: 3, maxLength: 1000 } }, additionalProperties: false }
    LandlordOnboardingApplication: { type: object, required: [id, applicantName, proposedOrganizationName, status, submittedAt, decidedAt, decisionReason], properties: { id: { type: string, format: uuid }, applicantName: { type: string }, proposedOrganizationName: { type: string }, status: { type: string, enum: [pending, approved, rejected] }, submittedAt: { type: string, format: date-time }, decidedAt: { type: [string, 'null'], format: date-time }, decisionReason: { type: [string, 'null'] } }, additionalProperties: false }
    LandlordOnboardingApplicationEnvelope: { type: object, required: [data, meta], properties: { data: { $ref: '#/components/schemas/LandlordOnboardingApplication' }, meta: { $ref: '#/components/schemas/Meta' } }, additionalProperties: false }
    LandlordOnboardingApplicationListEnvelope: { type: object, required: [items, meta], properties: { items: { type: array, items: { $ref: '#/components/schemas/LandlordOnboardingApplication' } }, meta: { $ref: '#/components/schemas/Meta' } }, additionalProperties: false }
    Envelope: { type: object, required: [data, auditEventId], properties: { data: { type: object, additionalProperties: true }, auditEventId: { type: string } } }
...
    PropertyInput: { type: object, required: [name, address, city], properties: { name: { type: string }, address: { type: string }, city: { type: string }, managerId: { type: string } } }
...